refactor(help-gen): rename main.swift to CLI.swift with @main entry-point
